+++ maintenance/projects/js-themes-toolkit-v8-x/packages/liferay-theme-tasks/lib/upgrade/6.2/upgrade.js
@@ -73,7 +73,7 @@ module.exports = function (options) {
 		lfrThemeConfig.removeDependencies(['liferay-theme-deps-6.2']);
 		lfrThemeConfig.setDependencies(
 			{
-				'liferay-theme-deps-7.0': '8.2.3',
+				'liferay-theme-deps-7.0': '8.2.4',
 			},
 			true
 		);